refac: discussions to features
